Make the deployer more generic
I was making assumptions about variables that would be present/required
in the app deploy playbook based on how we are deploying
Bocoup.com. Those assumptions would limit the way playbooks could be
written.

Instead of using a strict structure for the `ansible` object in the
message, I changed it to use a `BTreeMap<String, String>` so it can be
an arbitrary set of key and (string) values. This allows more
flexibility in the way users can design their app playbooks and makes
the deployer more generically useful.

There's a remaining limitation with this setup: because the `ansible`
key is expected to be BTreeMap<String, String>, any other type of value
will fail:

```json
{
  "secret": "...",
  "ansible": {
    "complex": {
      "will_fail": true
    },
    "boolean": false,
    "array": [1,2,3,4],
    "number": 100,
  }
}
```

Each one of those items under `ansible` will cause an error. I'm going
to explore a way to make it more generic, but I don't consider this a
huge limitation – ansible is pretty good at coercion, so as far as I can
tell its only complex structures (arrays, objects) that are really
affected.