Fix linux build script
`synchronize` pull wasn't working exactly as I expected: I was trying to
have it pull in a remote file and store it at a local location *under a
new name*. This works fine if the file already exists but fails if the
file doesn't exist.

Instead of trying to rename during the synchronize step, I've changed
the script to rename the file on the remote system before pulling it.