Fix git tests
SO it's apparently fucking impossible to add a folder that represents a
git repo without git trying to turn it into a submodule, which doesn't
work for me. Instead I'm gonna just tar up that folder and have in the
instructions that `make test` is the way to run the test suite, and as a
pre-test task I will untar the folder and put it in the place the suite
expects.