Refactor error reporting in repo_config.
Use an error enum that implements the error::Error trait. At the moment
I think this is actually slightly *worse* than what I had before as far
as outward error messaging (I lost the "subject" field, so it might be
harder to track down which branch config is causing an error) but I
think it's a better implementation to work from in general, and I'm
gonna work on figuring out how to get that back.