Implement queue limit option
This allows the administrator to set a hard queue limit in the hookshot
server configuration. Any new items that get added to the queue after
the limit has been reached bump off the oldest task scheduled.

For example, if we have a queue limit of 5 and a queue that looks like this:

```
<---
1 2 3 4 5
```

when a new task gets added the queue will look like this:

```
<---
2 3 4 5 6
```

If a task gets bumped from the queue the `cancel()` method will called
on that task.